Switch to GH actions for build

+++ build-package.ps1
@@ -1,25 +1,13 @@
-#Requires -Version 6.2
-#Requires -Modules @{ ModuleName = 'Microsoft.PowerShell.Archive'; ModuleVersion = '1.2.3' }
+#Requires -Version 5.1
 
 [CmdletBinding()]
 param (
   [string] $Configuration = 'Release',
-  [switch] $WithBinLog = $false
+  [switch] $ContinuousIntegration,
+  [switch] $WithBinLog
 )
 
-$opts = @( '--nologo' )
-if ($WithBinLog) {
-  $opts += '-bl'
-}
-
 $ErrorActionPreference = 'Stop'
-
-# The top-level folder used as target for the publish step (before being zipped).
-$PublishFolder = 'publish'
-
-# The zip file created after the publish step.
-# FIXME: Can we easily get any version info included in this name?
-$ZipFile = 'Zastai.NuGet.Server.zip'
 
 function Complete-BuildStep {
   param (
@@ -38,6 +26,29 @@ function Complete-BuildStep {
   }
 }
 
+$opts = @( '--nologo' )
+if ($WithBinLog) {
+  $opts += '-bl'
+}
+
+$props = @()
+if ($ContinuousIntegration) {
+  $props += '-p:ContinuousIntegrationBuild=true'
+  $props += '-p:Deterministic=true'
+}
+
+# The top-level folder used as target for the publish step (before being zipped).
+if ($ContinuousIntegration) {
+  $PublishFolder = "gh-build-${Configuration}"
+}
+else {
+  $PublishFolder = 'publish'
+}
+
+# The zip file created after the publish step.
+# FIXME: Can we easily get any version info included in this name?
+$ZipFile = 'Zastai.NuGet.Server.zip'
+
 Write-Host 'Cleaning up existing build output...'
 Remove-Item -Recurse -Force */bin, */obj
 if (Test-Path $PublishFolder) {
@@ -50,7 +61,7 @@ dotnet restore $opts
 Complete-BuildStep 'restore' 'PACKAGE RESTORE'
 
 Write-Host "Building the solution (Configuration: $Configuration)..."
-dotnet build $opts --no-restore "-c:$Configuration" '-p:ContinuousIntegrationBuild=true' '-p:Deterministic=true'
+dotnet build $opts --no-restore "-c:$Configuration" $props
 Complete-BuildStep 'build' 'SOLUTION BUILD'
 
 Write-Host 'Running tests...'
@@ -58,15 +69,12 @@ dotnet test $opts --no-build "-c:$Configuration"
 Complete-BuildStep 'test' 'UNIT TESTS'
 
 Write-Host 'Publishing...'
-dotnet publish $opts Zastai.NuGet.Server --no-build "-c:$Configuration" "-o:$PublishFolder"
+dotnet publish $opts PackageImport       --no-build "-c:$Configuration" "-o:$PublishFolder/importer"
+dotnet publish $opts Zastai.NuGet.Server --no-build "-c:$Configuration" "-o:$PublishFolder/server"
 Complete-BuildStep 'publish' 'PUBLISH'
 
-Write-Host 'Creating Zip File...'
-Push-Location $PublishFolder
-try {
-  Compress-Archive -Force -Path * -DestinationPath "..\$ZipFile"
+if (-not $ContinuousIntegration) {
+  Write-Host 'Creating Zip File...'
+  Compress-Archive -Force -Path $PublishFolder/* -DestinationPath $ZipFile
+  Remove-Item -Recurse -Force $PublishFolder
 }
-finally {
-  Pop-Location
-}
-Remove-Item -Recurse -Force $PublishFolder
